+++ server/src/main/java/com/shellwe/server/auth/handler/jwt/VerificationAuthFailedHandler.java
@@ -1,4 +1,4 @@
-package com.shellwe.server.auth.handler;
+package com.shellwe.server.auth.handler.jwt;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
@@ -11,7 +11,7 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
 @Slf4j
-public class verificationAuthFailedHandler implements AccessDeniedHandler {
+public class VerificationAuthFailedHandler implements AccessDeniedHandler {
 
     @Override
     public void handle(HttpServletRequest request, HttpServletResponse response, AccessDeniedException accessDeniedException) throws IOException, ServletException {
